requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
requests==2.24.0
selenium==3.141.0
soupsieve==2.0.1
urllib3==1.25.10
tkmacos==0.1.5